requirements.txt
streamlit
flask
deepface
opencv-python
pillow
numpy
gunicorn
requests
tf-keras